+++ manager_utils/tests/models.py
@@ -1,3 +1,4 @@
+from django.contrib.postgres.fields import JSONField, ArrayField
 from django.db import models
 from manager_utils import ManagerUtilsManager
 from timezone_field import TimeZoneField
@@ -11,6 +12,8 @@ class TestModel(models.Model):
     int_field = models.IntegerField(null=True, unique=True)
     char_field = models.CharField(max_length=128, null=True)
     float_field = models.FloatField(null=True)
+    json_field = JSONField(default=dict)
+    array_field = ArrayField(models.CharField(max_length=128), default=list)
     time_zone = TimeZoneField(default='UTC')
     json_data = JSONField(default=dict)