+++ run_tests.py
@@ -13,7 +13,7 @@ configure_settings()
 from django_nose import NoseTestSuiteRunner
 
 
-def run_tests(*test_args, **kwargs):
+def run(*test_args, **kwargs):
     if not test_args:
         test_args = ['manager_utils']
 
@@ -30,4 +30,4 @@ if __name__ == '__main__':
     parser.add_option('--verbosity', dest='verbosity', action='store', default=1, type=int)
     (options, args) = parser.parse_args()
 
-    run_tests(*args, **options.__dict__)
+    run(*args, **options.__dict__)