+++ docs/release_notes.rst
@@ -1,6 +1,15 @@
 Release Notes
 =============
 
+v1.3.2
+------
+* Fix bulk_update not properly casting fields
+* Add support for returning upserts with multiple unique fields for non native
+
+v1.3.1
+------
+* BAD RELEASE DO NOT USE
+
 v1.3.0
 ------
 * Updated version 2 interface